573: Redesign data sheet templates

+++ src/features/data_sheets/components/FormDataSheet.vue
@@ -19,7 +19,6 @@ const props = defineProps<{ record: Sheet; query: () => void }>();
 const { record } = toRefs(props);
 
 const entries = ref<Sheet["entries"]>({});
-const nonFieldErrors = ref<string[]>([]);
 const errors = ref<{ [key: string]: string[] }>({});
 
 const fields = computed(() => {
@@ -48,11 +47,11 @@ function getAttrs(uuid: string) {
 
 function change(field: SheetField, value: SheetValue) {
   errors.value = {};
-  const entries_include = Object.keys(entries.value).includes(field.uuid);
+  const entryExists = Object.keys(entries.value).includes(field.uuid);
   if (field.type === "file") {
     createFileEntry(field, value);
-  } else if (entries_include && !value) {
-    deleteEntry(field);
+  } else if (!value) {
+    if (entryExists) deleteEntry(field);
   } else {
     createOrUpdateEntry(field, value);
   }
@@ -66,7 +65,7 @@ function createFileEntry(field: SheetField, value: SheetValue) {
   formData.append("record_id", record.value.id.toString());
   formData.append("file", value as File);
   formData.append("action", "data_sheets/create_file_entry");
-  commandRequest(formData).catch((e) => handleError(field, e));
+  commandRequest(formData).catch((error) => handleError(field, error));
   record.value.entries[field.uuid] = (value as File).name;
 }
 
@@ -77,7 +76,7 @@ function createOrUpdateEntry(field: SheetField, value: SheetValue) {
     value: value,
     action: "data_sheets/create_or_update_entry",
   };
-  commandRequest(data).catch((e) => handleError(field, e));
+  commandRequest(data).catch((error) => handleError(field, error));
   record.value.entries[field.uuid] = value;
 }
 
@@ -87,7 +86,7 @@ function deleteEntry(field: SheetField) {
     record_id: record.value.id,
     action: "data_sheets/delete_entry",
   };
-  commandRequest(data).catch((e) => handleError(field, e));
+  commandRequest(data).catch((error) => handleError(field, error));
   delete record.value.entries[field.uuid];
 }
 
@@ -101,17 +100,8 @@ function handleError(field: SheetField, error: types.ICommandError) {
 </script>
 
 <template>
-  <form ref="form" novalidate @submit.prevent="">
+  <form novalidate @submit.prevent="">
     <div class="grid grid-cols-1 gap-4">
-      <div v-if="nonFieldErrors && nonFieldErrors.length">
-        <p
-          v-for="(error, index) in nonFieldErrors"
-          :key="index"
-          class="mt-2 text-sm leading-tight text-red-700"
-        >
-          {{ error }}
-        </p>
-      </div>
       <div
         v-for="field in fields"
         :key="field.name"
@@ -126,7 +116,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
           v-bind="getAttrs(field.uuid)"
           :label="field.label"
           :name="field.name"
-          required
+          :required="field.is_required"
           @change:model-value="change(field, $event)"
         />
         <FormSelect
@@ -134,7 +124,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
           v-bind="getAttrs(field.uuid)"
           :label="field.label"
           :name="field.name"
-          required
+          :required="field.is_required"
           :options="field.options"
           @update:model-value="change(field, $event)"
         />
@@ -143,7 +133,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
           v-bind="getAttrs(field.uuid)"
           :label="field.label"
           :name="field.name"
-          required
+          :required="field.is_required"
           :options="field.options"
           @update:model-value="change(field, $event)"
         />
@@ -153,7 +143,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
           v-bind="getAttrs(field.uuid)"
           :label="field.label"
           :name="field.name"
-          required
+          :required="field.is_required"
           @update:model-value="change(field, $event)"
         >
           <ButtonNormal
@@ -165,7 +155,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
             Download
           </ButtonNormal>
           <ButtonNormal
-            v-if="field.name in entries"
+            v-if="field.uuid in entries"
             kind="delete"
             class="hover:text-opacity-75 rounded font-medium text-gray-700 focus:outline-none"
             @click="change(field, '')"
@@ -179,7 +169,7 @@ function handleError(field: SheetField, error: types.ICommandError) {
           v-bind="getAttrs(field.uuid)"
           :label="field.label"
           :type="field.type"
-          required
+          :required="field.is_required"
           @change:model-value="change(field, $event)"
         />
         <p
